+++ src/lib/cnd.h
@@ -11,13 +11,13 @@ void r_interrupt();
 
 sexp* r_interp_str(const char* fmt, ...);
 
-sexp* r_new_condition(sexp* type, sexp* data, sexp* msg);
+sexp* r_new_condition(sexp* type, sexp* msg, sexp* call, sexp* data);
 
 static inline bool r_is_condition(sexp* x) {
   return TYPEOF(x) == VECSXP && Rf_inherits(x, "condition");
 }
 
-void r_cnd_signal(sexp* cnd, bool mufflable);
+void r_cnd_signal(sexp* cnd);
 void r_cnd_inform(sexp* cnd, bool mufflable);
 void r_cnd_warn(sexp* cnd, bool mufflable);
 void r_cnd_abort(sexp* cnd, bool mufflable);