+++ src/lib/vec-chr.c
@@ -94,3 +94,11 @@ sexp* chr_append(sexp* chr, sexp* r_str) {
   FREE(1);
   return out;
 }
+
+
+sexp* r_shared_empty_chr = NULL;
+
+void r_init_library_vec_chr() {
+  r_shared_empty_chr = r_scalar_chr("");
+  r_mark_precious(r_shared_empty_chr);
+}